+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Engelsystem changelog
 
-## NEXT
+## [v.7.1.0](https://github.com/johnjohndoe/engelsystem/releases/tag/v.7.1.0)
 
-* Published: Not released yet.
+* Published: 2022-05-20
 
 ### Changes